requirements-dev.txt
-r requirements.txt
coverage[toml]~=7.6.1
mypy~=1.15.0
pytest~=7.2.1
pytest-cov~=5.0.0
requests-mock~=1.12.1
ruff~=0.11.5
setuptools~=75.1.0
setuptools-scm[toml]~=8.1.0
tox~=4.25.0
tox-uv~=1.25.0
types-requests~=2.31.0
types-psutil~=7.0.0.20250218
pandas-stubs~=2.2.2
pyarrow-stubs~=20.0.0
requests-mock~=1.12.1
click~=8.1.7
checksumdir~=1.2.0